+++ semaphoreci.sh
@@ -8,7 +8,12 @@ setup() {
 	then
 		curl -L https://get.haskellstack.org/stable/linux-x86_64.tar.gz | tar xz --wildcards --strip-components=1 -C $SEMAPHORE_CACHE_DIR '*/stack'
 	fi
+	if [ -e "$SEMAPHORE_CACHE_DIR/graphql.cabal" ]
+	then
+            cp -a $SEMAPHORE_CACHE_DIR/graphql.cabal graphql.cabal
+        fi
 	$STACK --no-terminal setup
+        cp -a graphql.cabal $SEMAPHORE_CACHE_DIR/graphql.cabal
 }
 
 setup_test() {